requirements.txt
aioredis==1.3.1
asgiref==3.2.10
async-timeout==3.0.1
attrs==19.3.0
autobahn==20.6.1
Automat==20.2.0
beautifulsoup4==4.9.1
certifi==2020.6.20
cffi==1.14.0
channels==2.4.0
channels-redis==2.4.0
chardet==3.0.4
constantly==15.1.0
cryptography==2.9.2
daphne==2.5.0
defusedxml==0.7.0rc1
Django==3.0.7
django-bootstrap3==14.0.0
django-bootstrap4==2.1.1
django-crispy-forms==1.9.1
django-extensions==2.2.9
django-taggit==1.2.0
django-verified-email-field==1.6.1
hiredis==1.0.1
hyperlink==19.0.0
idna==2.9
image==1.5.32
importlib-metadata==1.6.1
incremental==17.5.0
msgpack==0.6.2
oauthlib==3.1.0
Pillow==7.1.2
psycopg2==2.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
PyJWT==1.7.1
pyOpenSSL==19.1.0
python-decouple==3.3
python3-openid==3.1.0
pytz==2020.1
redis==3.4.1
requests==2.24.0
requests-oauthlib==1.3.0
service-identity==18.1.0
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
soupsieve==2.0.1
sqlparse==0.3.1
Twisted==20.3.0
txaio==20.4.1
urllib3==1.25.9
zipp==3.1.0
zope.interface==5.1.0
